requirements.txt
gradio
requests
python-dotenv
python-docx
PyPDF2